Parametrize phase correlation low pass filter sigma
The phase-correlation image is low-pass filtered by multiplying the
cross-power spectrum with Gaussian weights. The sigma parameter of the
Gaussian function varies depending on the input images, and must be
parametrizable. So far it was hardcoded based on image size, but now
`find_translation` takes it in as a parameter. A good value for my
images was chosen as the default value, until proper parametrization is
added.