Fix rotation problem in image_generator.py
The problem was that the input image wasn't square, which leads to the
FFT not being square, which means circles in the amplitude spectrum end
up being ovals, which don't get projected to vertical lines in polar
coordinates, but into waves instead. This is now fixed in the
image_generator.py script, which was very helpful in troubleshooting
this.

Next it must be fixed in the actual code.