Adjust circle masking parameters
The inner radius of the circle mask was increased a bit to cover the
very bright prominences in the 2024 eclipse.

The image is now also roughly stretched before converting to 8-bit for
circle finding, as the images with shortest exposure were almost black
after converting to 8-bit, and the circle finding was struggling.