Extract CLI commands to separate files
The main.py file was starting to get messy with all the commands and
their dependencies. Now each command is in its own module, and they are
explicitly added to the Click group. For some reason PyCharm didn't
understand the types, so the types are ignored for the add_command
calls.