Implement max mask mode
The max mode means both the alignment and preprocess-only commands go
through all the images to find the biggest mask that is enough to cover
the moon and any overexposed areas in all the given photos, and it will
use that mask size to preprocess all images.

In the regular auto mode it will find a different mask size for each
image.

The constant mask size helps with aligning very overexposed images to a
reference that is not overexposed.