Improve linear fitting
Instead of having hardcoded lower limits for the data points, the lower
limits are separate for x and y, and are based on the 80th percentile of
the data to ignore the noisy bottom parts of the data that skews the
slope of the linear fit. We also filter the amount of data used in
fitting to improve performance. We don't need to use 50 million data
points to fit a line, when 10 000 will do.