Use better moon masking in linear fitting
Now the moon is detected in each image, and we only use pixels that are
not contaminated by the moon in neither the reference image or the image
being fitted. We need to anyway find the moon in each image to be able
to handle the image weights correctly for the moon area, so this comes
at no extra cost.