Rewrite HDR stacking
Instead of linear fitting all images to a reference frame, the stacking
now sorts the images into brightness order, fits together adjacent
images in the sorted order, and finally calculates a global fit to the
reference image using least squares.

To prevent breaking of prominences and other highlights, the highlights
are no longer set to have weight 0.0 for the reference frame. For other
frames it works as before.

The RANSAC and circle finding parameters were also tuned for better
accuracy.