Implement unsharp masking with partial convolution and inpainting
It is now possible to sharpen coronal details with unsharp masking that
allows you to mask out the moon and other non-coronal features. The
filter uses partial convolution and an adaptive convolution kernel that
varies spatially according to distance and direction from the moon/sun
center point. Currently it does not really differ from a Gaussian
kernel, but will allow you to adjust the tangential and radial
parameters separately.

Regular partial convolution will still introduce some edge artifacts
when mask edges are close to gradients perpendicular to the mask edge.
The edge artifacts are milder than with regular convolution, though. To
reduce the artifacts dramatically, we inpaint masked pixels close to
mask edges, so we can include them in the convolution. The inpainting
takes the gradient of the neighbouring unmasked pixels into account to
reduce the bias introduced by the convolution.

The 32-bit floating point images produced by the filtering include the
sRGB ICC profile, despite not strictly being sRGB images. If you open
32-bit images without an embedded ICC profile in Adobe Photoshop,
Photoshop will apply a gamma correction to the images, which we do not
want. Embedding an ICC profile in the images works around this.