Simplify image weighting
Instead of the manually crafted sigmoid function thingy used for
weighting different exposures in the HDR composition, we now use a
quadratic "hat" function that is zero at 0.0 and 1.0, and non-zero
between them. This ensures no pixels end up empty because of the
weighting, which happened with the sigmoid-based function.

We also fill any unfilled pixels explicitly with data from the reference
image instead of using a hacky flag in the weighting function.