Fix color calibration with overexposed images
The color calibration was not working properly with overexposed images,
because it wasn't taking into account that the overexposure might happen
at different parts of the image for different color channels, leading to
a skew in the color calibration. Now it properly selects valid pixels
based on all the color channels.