Use Fourier interpolation to tangentially blur moon glare model
The smooth model is subtracted from the image to remove the glare. This
doesn't work perfectly yet. Without some radial smoothing with a
Gaussian filter there are artifacts close to the center of the moo,
while the radial smoothing messes with the moon limb brightness, causing
the correction to be incorrect.

I might replace this with a different approach.